move application into sub-folder
