+++ .pre-commit-config.yaml
@@ -36,7 +36,7 @@ repos:
     hooks:
       - id: pydocstyle
         additional_dependencies: ["pydocstyle[toml]"]
-        exclude: ^migrations/|__init__.py
+        exclude: ^app/migrations/|__init__.py
   - repo: https://github.com/PyCQA/bandit
     rev: 1.7.5
     hooks:
