+++ app/resources/home.py
@@ -4,8 +4,8 @@ from typing import Union
 from fastapi import APIRouter, Header, Request
 from fastapi.templating import Jinja2Templates
 
-from config.helpers import get_api_version
-from config.settings import get_settings
+from app.config.helpers import get_api_version
+from app.config.settings import get_settings
 
 router = APIRouter()
 templates = Jinja2Templates(directory="templates")
