+++ tests/unit/test_auth_manager.py
@@ -3,12 +3,12 @@ from datetime import datetime
 
 import jwt
 import pytest
-from fastapi import BackgroundTasks, HTTPException
-
 from config.settings import get_settings
+from fastapi import BackgroundTasks, HTTPException
 from managers.auth import AuthManager, ResponseMessages
 from managers.user import UserManager
 from schemas.request.auth import TokenRefreshRequest
+
 from tests.helpers import get_token
